pyproject.toml
[project]
name = "pocket-scion-osc"
version = "1.0.0"
description = "Pocket Scion OSC bridge"
requires-python = ">=3.13"
dependencies = ["python-rtmidi", "python-osc"]
